horsebreed now alters the Horse's stats.  - reducedFallRate alters the rate the horse falls  - breedSpeed alters the pace of the horse
 Note: distanceTravelled is now a double which is converted into an int during the race to express minute differences in distance

+++ Part2/RaceGUI.java
@@ -79,7 +79,6 @@ public class RaceGUI {
             horse.rise();
         }
 
-        ;
         while (!race.getFinished())
         {
             //move each horse
@@ -128,7 +127,7 @@ public class RaceGUI {
     {
         for (int i = 0; i < race.getLaneCount(); i++)
         {
-            int xPosition = (int) (TRACK_DISTANCE / (double) race.getRaceLength()) * race.getHorses().get(i).getDistanceTravelled();
+            int xPosition = (int) (TRACK_DISTANCE / (double) race.getRaceLength()) * (int) race.getHorses().get(i).getDistanceTravelled();
             int yPosition = horseLabels.get(i).getLocation().y;
             horseLabels.get(i).setLocation(xPosition, yPosition);
             horseLabels.get(i).repaint();
